+++ sdk/core/Azure.Core/src/HttpMessage.cs
@@ -87,16 +87,12 @@ namespace Azure.Core
 
         internal int RetryNumber { get; set; }
 
-        internal int ResponseNumber { get; set; }
-
-        internal Exception? LastException { get; set; }
-
         internal DateTimeOffset ProcessingStartTime { get; set; }
 
         /// <summary>
         /// The processing context for the message.
         /// </summary>
-        public ProcessingContext ProcessingContext => new(this);
+        internal ProcessingContext ProcessingContext => new(this);
 
         internal void ApplyRequestContext(RequestContext? context, ResponseClassifier? classifier)
         {