+++ sdk/keyvault/Azure.Security.KeyVault.Keys/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.5.0-beta.1 (Unreleased)
+## 4.5.0-beta.1 (2022-11-09)
 
 ### Features Added
 
@@ -9,8 +9,6 @@
 - Added `KeyType.Okp` and `KeyType.OkpHsm` for `JsonWebKey`.
 - Added `KeyCurveName.Ed25519` to create an Octet Key Pair (OKP) using the Ed25519 curve.
 - Added `SignatureAlgorithm.EdDsa` to support signing and verifying using an Edwards Curve Digital Signature Algorithm (EdDSA) on Managed HSM.
-
-### Breaking Changes
 
 ### Bugs Fixed